pyproject.toml
[project]
name = "uicdev"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0.3",
    "django-jazzmin>=3.0.4",
    "djangorestframework>=3.17.0",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular>=0.29.0",
    "pre-commit>=4.5.1",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.2",
    "rest-framework-simplejwt>=0.0.2",
    "ruff>=0.15.7",
]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.ruff.lint.isort]
known-first-party = ["apps", "config"]
